# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Vaeloom Backend
  version: 0.2.0
paths:
...
  /api/v1/workspaces/{workspace_id}/agents:
    get:
      tags:
        - workspaces
      summary: List Workspace Agents
      operationId: list_workspace_agents_api_v1_workspaces__workspace_id__agents_get
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/agents:
    post:
      tags:
        - agents
      summary: Create Agent
      operationId: create_agent_api_v1_agents_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - agents
      summary: List Agents
      operationId: list_agents_api_v1_agents_get
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: category
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Category
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Agents Api V1 Agents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/chat:
    post:
      tags:
        - agents
      summary: Chat
      description: 'High-level chat endpoint: auto-classifies intent, routes to the
        right agent,

        runs the agentic loop, and returns the result. If dto.agentName is provided

        (enterprise explicit routing), honors it with high confidence.'
      operationId: chat_api_v1_agents_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessage'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/{agent_id}:
    get:
      tags:
        - agents
      summary: Get Agent
      operationId: get_agent_api_v1_agents__agent_id__get
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
        - agents
      summary: Update Agent
      operationId: update_agent_api_v1_agents__agent_id__put
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - agents
      summary: Deactivate Agent
      operationId: deactivate_agent_api_v1_agents__agent_id__delete
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/{agent_id}/execute:
    post:
      tags:
        - agents
      summary: Execute Agent
      operationId: execute_agent_api_v1_agents__agent_id__execute_post
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecute'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/{agent_id}/executions:
    get:
      tags:
        - agents
      summary: List Executions
      operationId: list_executions_api_v1_agents__agent_id__executions_get
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Page Size
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Executions Api V1 Agents  Agent Id  Executions
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/{agent_id}/run:
    post:
      tags:
        - agents
      summary: Run Agent
      operationId: run_agent_api_v1_agents__agent_id__run_post
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecute'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/agents/{agent_id}/schedule:
    post:
      tags:
        - agents
      summary: Schedule Agent
      operationId: schedule_agent_api_v1_agents__agent_id__schedule_post
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Agent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AgentCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        category:
          type: string
          maxLength: 100
          minLength: 1
          title: Category
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        config:
          additionalProperties: true
          type: object
          title: Config
      type: object
      required:
        - name
        - category
      title: AgentCreate
    AgentExecute:
      properties:
        input:
          additionalProperties: true
          type: object
          title: Input
        stream:
          type: boolean
          title: Stream
          default: false
      type: object
      title: AgentExecute
    AgentResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
        status:
          type: string
          title: Status
        version:
          type: string
          title: Version
        config:
          additionalProperties: true
          type: object
          title: Config
        capabilities:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Capabilities
        permissions:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Permissions
        workspace_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Workspace Id
        user_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: User Id
        tenant_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Tenant Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - name
        - status
        - version
        - config
        - created_at
        - updated_at
      title: AgentResponse
    AgentUpdate:
      properties:
        name:
          anyOf:
            - type: string
            - type: 'null'
          title: Name
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        config:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Config
        status:
          anyOf:
            - type: string
            - type: 'null'
          title: Status
      type: object
      title: AgentUpdate
...
    ChatMessage:
      properties:
        workspaceId:
          type: string
          title: Workspaceid
        message:
          type: string
          title: Message
        agentName:
          anyOf:
            - type: string
            - type: 'null'
          title: Agentname
      type: object
      required:
        - workspaceId
        - message
      title: ChatMessage
...
    ExecutionResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        agent_id:
          type: string
          format: uuid
          title: Agent Id
        status:
          type: string
          title: Status
        input:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Input
        output:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Output
        error:
          anyOf:
            - type: string
            - type: 'null'
          title: Error
        tokens_used:
          anyOf:
            - type: integer
            - type: 'null'
          title: Tokens Used
        cost:
          anyOf:
            - type: number
            - type: 'null'
          title: Cost
        duration_ms:
          anyOf:
            - type: integer
            - type: 'null'
          title: Duration Ms
        duration:
          anyOf:
            - type: integer
            - type: 'null'
          title: Duration
        started_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Started At
        completed_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Completed At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
        - id
        - agent_id
        - status
        - created_at
      title: ExecutionResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ScheduleRequest:
      properties:
        cron:
          type: string
          minLength: 1
          title: Cron
        input:
          additionalProperties: true
          type: object
          title: Input
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
        - cron
      title: ScheduleRequest
    ScheduleResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        agent_id:
          type: string
          format: uuid
          title: Agent Id
        cron:
          type: string
          title: Cron
        input:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Input
        enabled:
          type: boolean
          title: Enabled
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
        - id
        - agent_id
        - cron
        - enabled
        - created_at
        - updated_at
      title: ScheduleResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...